2025-11-03 vartab_lock() athelyezve oref_new/vref_new-ba; PUSH assign_lock vedelem alatt; operator= helyett valuecopy_lk; valuemove() helyett valuecopy_lk; assign()-ban valuecopy_lk;

+++ usr/include/varlock.h
@@ -25,7 +25,7 @@ struct varlock
 
     int lock(void *p)
     {
-        int x=((unsigned long)p)%prime;
+        int x=((unsigned long long)p)%prime;
         pthread_mutex_lock( &mutex[x] );
         return x;
     };
@@ -38,7 +38,7 @@ struct varlock
 
     int lock_try(void *p)
     {
-        int x=((unsigned long)p)%prime;
+        int x=((unsigned long long)p)%prime;
         if( pthread_mutex_trylock(&mutex[x]) ) // -> 0 vagy errno
         {
             return -1;
